Toggle Player Width
Added an expand/collapse button to allow the video to be larger for better viewing.

Found the video player too small when watching a video at work, fullscreen is not an option for me as I like to use other things on my screen whilst watching a video. This will avoid that issue!